levis (webportal)/requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.18
certifi==2018.10.15
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
defusedxml==0.5.0
Django==2.1.5
django-allauth==0.38.0
django-celery==3.3.1
django-cors-middleware==1.3.1
django-mysql==2.4.1
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.5
django-saml2-auth==2.2.1
django-session-timeout==0.0.4
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
idna==2.7
importlib-metadata==0.23
kombu==3.0.37
more-itertools==7.2.0
numpy==1.16.4
oauthlib==3.0.0
pandas==0.24.2
pbr==5.4.3
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pysaml2==4.9.0
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.3
redis==2.10.3
requests==2.20.1
requests-oauthlib==1.2.0
six==1.12.0
stevedore==1.31.0
urllib3==1.24.1
vine==1.3.0
virtualenv==16.7.7
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
xlrd==1.2.0
XlsxWriter==1.1.8
zipp==0.6.0
